OneDrive/Desktop/Github Projects/email-whatsapp-agent/requirements.txt
google-api-python-client==2.99.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
openai==1.3.0
twilio==8.13.0
python-dotenv==1.0.0
schedule==1.2.0
